# artifacts/api-server/src/routes/documents.ts
import { Router } from "express";
import { db, usersTable, documentsTable, aiDraftsTable } from "@workspace/db";
import { eq, desc, and, sql } from "drizzle-orm";
import { requireAuth } from "../lib/auth";
import { CreateDocumentBody, UpdateDocumentBody, ListDocumentsQueryParams } from "@workspace/api-zod";
import { DOCUMENT_TYPES, PLAN_LIMITS, generateDocumentContent } from "../lib/document-templates";
import { aiService } from "../lib/ai";

const router = Router();

router.get("/document-types", async (_req, res): Promise<void> => {
  res.json(DOCUMENT_TYPES);
});

router.get("/documents/dashboard-stats", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const allDocs = await db.select().from(documentsTable).where(eq(documentsTable.userId, user[0].id));

  const now = new Date();
  const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);
  const docsThisMonth = allDocs.filter(d => new Date(d.createdAt) >= startOfMonth);

  const byType: Record<string, number> = {};
  for (const doc of allDocs) {
    byType[doc.type] = (byType[doc.type] || 0) + 1;
  }

  const planInfo = PLAN_LIMITS[user[0].plan] || PLAN_LIMITS.free;

  res.json({
    totalDocuments: allDocs.length,
    documentsThisMonth: docsThisMonth.length,
    documentsByType: Object.entries(byType).map(([type, count]) => ({ type, count })),
    planLimits: {
      plan: user[0].plan,
      limit: planInfo.documentsPerMonth,
      used: docsThisMonth.length,
    },
  });
});

router.get("/documents/recent", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const docs = await db.select().from(documentsTable)
    .where(eq(documentsTable.userId, user[0].id))
    .orderBy(desc(documentsTable.createdAt))
    .limit(5);

  res.json(docs.map(d => ({
    id: d.id,
    userId: d.userId,
    title: d.title,
    type: d.type,
    status: d.status,
    content: d.content,
    metadata: d.metadata,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
  })));
});

router.get("/documents", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const queryParsed = ListDocumentsQueryParams.safeParse(req.query);
  const filters: any[] = [eq(documentsTable.userId, user[0].id)];

  if (queryParsed.success) {
    if (queryParsed.data.type) {
      filters.push(eq(documentsTable.type, queryParsed.data.type));
    }
    if (queryParsed.data.status) {
      filters.push(eq(documentsTable.status, queryParsed.data.status));
    }
  }

  const docs = await db.select().from(documentsTable)
    .where(and(...filters))
    .orderBy(desc(documentsTable.createdAt));

  res.json(docs.map(d => ({
    id: d.id,
    userId: d.userId,
    title: d.title,
    type: d.type,
    status: d.status,
    content: d.content,
    metadata: d.metadata,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
  })));
});

router.post("/documents", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const parsed = CreateDocumentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { title, type, answers } = parsed.data;

  const planInfo = PLAN_LIMITS[user[0].plan] || PLAN_LIMITS.free;
  if (!planInfo.documentTypes.includes(type)) {
    res.status(403).json({ error: "This document type requires a higher plan" });
    return;
  }

  if (planInfo.documentsPerMonth !== null) {
    const now = new Date();
    const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);
    const countResult = await db.select({ count: sql<number>`count(*)::int` })
      .from(documentsTable)
      .where(and(
        eq(documentsTable.userId, user[0].id),
        sql`${documentsTable.createdAt} >= ${startOfMonth}`
      ));
    const usedThisMonth = Number(countResult[0]?.count ?? 0);
    if (usedThisMonth >= planInfo.documentsPerMonth) {
      res.status(403).json({ error: "Monthly document limit reached. Please upgrade your plan." });
      return;
    }
  }

  // Convert answers to a readable string for the prompt
  const answersText = Object.entries(answers || {})
    .map(([k, v]) => `- ${k.replace(/_/g, " ")}: ${v}`)
    .join("\n");

  const prompt = `Generate a comprehensive, legally sound ${type.replace(/_/g, " ")}. 
The document must strictly follow a structured Markdown format containing these exact sections where applicable:
# [Document Title]
## Parties
## Effective Date
## Definitions
## Obligations & Clauses
## Termination
## Dispute Resolution
## Governing Law
## Signatures

Use professional Indian legal terminology.

User Requirements:
${answersText}`;

  // Call the AI Service
  let content = "";
  try {
    const draftResult = await aiService.generateDraft(prompt, "You are an expert Indian corporate lawyer and contract drafter.");
    content = draftResult.content;
  } catch (err: any) {
    console.error("AI Generation Error:", err);
    res.status(500).json({ error: "Failed to generate document via AI." });
    return;
  }

  const doc = await db.insert(documentsTable).values({
    userId: user[0].id,
    title,
    type,
    status: "completed",
    content,
    metadata: JSON.stringify(answers),
  }).returning();

  const d = doc[0];
  res.status(201).json({
    id: d.id,
    userId: d.userId,
    title: d.title,
    type: d.type,
    status: d.status,
    content: d.content,
    metadata: d.metadata,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
  });
});

router.get("/documents/:id", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) {
    res.status(400).json({ error: "Invalid ID" });
    return;
  }

  const doc = await db.select().from(documentsTable)
    .where(and(eq(documentsTable.id, id), eq(documentsTable.userId, user[0].id)))
    .limit(1);

  if (!doc.length) {
    res.status(404).json({ error: "Document not found" });
    return;
  }

  const d = doc[0];
  res.json({
    id: d.id,
    userId: d.userId,
    title: d.title,
    type: d.type,
    status: d.status,
    content: d.content,
    metadata: d.metadata,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
  });
});

router.patch("/documents/:id", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) {
    res.status(400).json({ error: "Invalid ID" });
    return;
  }

  const parsed = UpdateDocumentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const updated = await db.update(documentsTable)
    .set(parsed.data)
    .where(and(eq(documentsTable.id, id), eq(documentsTable.userId, user[0].id)))
    .returning();

  if (!updated.length) {
    res.status(404).json({ error: "Document not found" });
    return;
  }

  const d = updated[0];
  res.json({
    id: d.id,
    userId: d.userId,
    title: d.title,
    type: d.type,
    status: d.status,
    content: d.content,
    metadata: d.metadata,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
  });
});

router.delete("/documents/:id", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) {
    res.status(400).json({ error: "Invalid ID" });
    return;
  }

  await db.delete(documentsTable)
    .where(and(eq(documentsTable.id, id), eq(documentsTable.userId, user[0].id)));

  res.status(204).send();
});

router.get("/documents/:id/download", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) {
    res.status(400).json({ error: "Invalid ID" });
    return;
  }

  const doc = await db.select().from(documentsTable)
    .where(and(eq(documentsTable.id, id), eq(documentsTable.userId, user[0].id)))
    .limit(1);

  if (!doc.length) {
    res.status(404).json({ error: "Document not found" });
    return;
  }

  const d = doc[0];
  res.json({
    id: d.id,
    title: d.title,
    content: d.content || "",
  });
});

// AI Draft generation
router.post("/documents/ai-draft", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const { prompt, documentId, context } = req.body;
  if (!prompt) {
    res.status(400).json({ error: "Prompt required" });
    return;
  }

  try {
    const result = await aiService.generateDraft(prompt, context);

    // Save draft to database
    const [draft] = await db.insert(aiDraftsTable).values({
      userId: user[0].id,
      documentId: documentId ? parseInt(documentId) : null,
      prompt,
      generatedContent: result.content,
      model: "gpt-4o",
      tokensUsed: result.tokensUsed,
    }).returning();

    res.status(201).json({
      id: draft.id,
      content: result.content,
      tokensUsed: result.tokensUsed,
    });
  } catch (error) {
    console.error("AI draft error:", error);
    res.status(500).json({ error: "Failed to generate AI draft" });
  }
});

// Get AI drafts for user
router.get("/ai-drafts", requireAuth, async (req, res): Promise<void> => {
  const clerkUserId = (req as any).clerkUserId as string;
  const user = await db.select().from(usersTable).where(eq(usersTable.clerkId, clerkUserId)).limit(1);
  if (!user.length) {
    res.status(401).json({ error: "User not found" });
    return;
  }

  const drafts = await db.select()
    .from(aiDraftsTable)
    .where(eq(aiDraftsTable.userId, user[0].id))
    .orderBy(desc(aiDraftsTable.createdAt))
    .limit(20);

  res.json(drafts);
});

export default router;