# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LexAI — Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfilePlan = typeof UserProfilePlan[keyof typeof UserProfilePlan];


export const UserProfilePlan = {
  free: 'free',
  pro: 'pro',
  enterprise: 'enterprise',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  plan: UserProfilePlan;
  createdAt: string;
}

export type SubscriptionPlanProperty = typeof SubscriptionPlanProperty[keyof typeof SubscriptionPlanProperty];


export const SubscriptionPlanProperty = {
  free: 'free',
  pro: 'pro',
  enterprise: 'enterprise',
} as const;

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  cancelled: 'cancelled',
  past_due: 'past_due',
} as const;

export interface Subscription {
  id: number;
  userId: number;
  plan: SubscriptionPlanProperty;
  status: SubscriptionStatus;
  /** @nullable */
  currentPeriodEnd: string | null;
  cancelAtPeriodEnd?: boolean;
}

export interface SubscriptionPlan {
  id: string;
  name: string;
  price: number;
  interval: string;
  features: string[];
  documentTypes: string[];
  /** @nullable */
  documentsPerMonth?: number | null;
}

export type SubscriptionUpgradeInputPlan = typeof SubscriptionUpgradeInputPlan[keyof typeof SubscriptionUpgradeInputPlan];


export const SubscriptionUpgradeInputPlan = {
  free: 'free',
  pro: 'pro',
  enterprise: 'enterprise',
} as const;

export interface SubscriptionUpgradeInput {
  plan: SubscriptionUpgradeInputPlan;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  draft: 'draft',
  completed: 'completed',
} as const;

export interface Document {
  id: number;
  userId: number;
  title: string;
  type: string;
  status: DocumentStatus;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type DocumentInputAnswers = { [key: string]: unknown };

export interface DocumentInput {
  title: string;
  type: string;
  answers: DocumentInputAnswers;
}

export type DocumentUpdateStatus = typeof DocumentUpdateStatus[keyof typeof DocumentUpdateStatus];


export const DocumentUpdateStatus = {
  draft: 'draft',
  completed: 'completed',
} as const;

export interface DocumentUpdate {
  title?: string;
  content?: string;
  status?: DocumentUpdateStatus;
}

export interface DocumentDownload {
  id: number;
  title: string;
  content: string;
}

export type DocumentTypeRequiredPlan = typeof DocumentTypeRequiredPlan[keyof typeof DocumentTypeRequiredPlan];


export const DocumentTypeRequiredPlan = {
  free: 'free',
  pro: 'pro',
  enterprise: 'enterprise',
} as const;

export type DocumentQuestionType = typeof DocumentQuestionType[keyof typeof DocumentQuestionType];


export const DocumentQuestionType = {
  text: 'text',
  textarea: 'textarea',
  date: 'date',
  select: 'select',
} as const;

export interface DocumentQuestion {
  id: string;
  label: string;
  type: DocumentQuestionType;
  options?: string[];
  required?: boolean;
  placeholder?: string;
}

export interface DocumentType {
  id: string;
  name: string;
  description: string;
  category: string;
  requiredPlan: DocumentTypeRequiredPlan;
  questions?: DocumentQuestion[];
}

export type DashboardStatsDocumentsByTypeItem = {
  type: string;
  count: number;
};

export type DashboardStatsPlanLimits = {
  plan: string;
  /** @nullable */
  limit: number | null;
  used: number;
};

export interface DashboardStats {
  totalDocuments: number;
  documentsThisMonth: number;
  documentsByType: DashboardStatsDocumentsByTypeItem[];
  planLimits: DashboardStatsPlanLimits;
}

export type ContractStatus = typeof ContractStatus[keyof typeof ContractStatus];


export const ContractStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Contract {
  id: number;
  userId: number;
  title: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  /** @nullable */
  content?: string | null;
  status: ContractStatus;
  uploadedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export type ContractUploadResponseStatus = typeof ContractUploadResponseStatus[keyof typeof ContractUploadResponseStatus];


export const ContractUploadResponseStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface ContractUploadResponse {
  contractId: number;
  status: ContractUploadResponseStatus;
}

export interface RAGQueryInput {
  question: string;
}

export interface RAGQueryResponse {
  answer: string;
}

export type ListDocumentsParams = {
type?: string;
status?: string;
};

export type UploadContractBody = {
  file: Blob;
  userId: string;
  title?: string;
};