# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LexAI — Legal Document Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Legal document operations
  - name: subscriptions
    description: Subscription management
  - name: users
    description: User profile operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  /users/me/subscription:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user subscription
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized

  /subscriptions/plans:
    get:
      operationId: getSubscriptionPlans
      tags: [subscriptions]
      summary: List available subscription plans
      responses:
        "200":
          description: Available plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionPlan"

  /subscriptions/upgrade:
    post:
      operationId: upgradeSubscription
      tags: [subscriptions]
      summary: Upgrade or change subscription plan (mock billing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpgradeInput"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized

  /subscriptions/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel current subscription
      responses:
        "200":
          description: Cancelled subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List user's generated documents
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
    post:
      operationId: createDocument
      tags: [documents]
      summary: Generate a new legal document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Generated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
        "403":
          description: Plan limit reached

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a specific document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Get document download URL or content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDownload"
        "404":
          description: Not found

  /documents/dashboard-stats:
    get:
      operationId: getDashboardStats
      tags: [documents]
      summary: Dashboard statistics for the current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized

  /documents/recent:
    get:
      operationId: getRecentDocuments
      tags: [documents]
      summary: Get recently generated documents (last 5)
      responses:
        "200":
          description: Recent documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized

  /document-types:
    get:
      operationId: getDocumentTypes
      tags: [documents]
      summary: Get available document types with plan access info
      responses:
        "200":
          description: Document types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentType"

  /contracts/upload:
    post:
      operationId: uploadContract
      tags: [contracts]
      summary: Upload a contract for analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, userId]
              properties:
                file:
                  type: string
                  format: binary
                userId:
                  type: string
                title:
                  type: string
      responses:
        "200":
          description: Upload response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractUploadResponse"
        "400":
          description: Bad request

  /contracts/user/{userId}:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all uploaded contracts for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found

  /contracts/{id}/query:
    post:
      operationId: queryContract
      tags: [contracts]
      summary: Query a contract using RAG
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RAGQueryInput"
      responses:
        "200":
          description: Query answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGQueryResponse"
        "400":
          description: Bad request
        "500":
          description: Server error

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    UserProfile:
      type: object
      required: [id, clerkId, email, plan, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro, enterprise]
        createdAt:
          type: string

    Subscription:
      type: object
      required: [id, userId, plan, status, currentPeriodEnd]
      properties:
        id:
          type: integer
        userId:
          type: integer
        plan:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, cancelled, past_due]
        currentPeriodEnd:
          type: ["string", "null"]
        cancelAtPeriodEnd:
          type: boolean

    SubscriptionPlan:
      type: object
      required: [id, name, price, interval, features, documentTypes]
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        interval:
          type: string
        features:
          type: array
          items:
            type: string
        documentTypes:
          type: array
          items:
            type: string
        documentsPerMonth:
          type: ["integer", "null"]

    SubscriptionUpgradeInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [free, pro, enterprise]

    Document:
      type: object
      required: [id, userId, title, type, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [draft, completed]
        content:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    DocumentInput:
      type: object
      required: [title, type, answers]
      properties:
        title:
          type: string
        type:
          type: string
        answers:
          type: object
          additionalProperties: true

    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, completed]

    DocumentDownload:
      type: object
      required: [id, title, content]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string

    DocumentType:
      type: object
      required: [id, name, description, category, requiredPlan]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        requiredPlan:
          type: string
          enum: [free, pro, enterprise]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/DocumentQuestion"

    DocumentQuestion:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, textarea, date, select]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
        placeholder:
          type: string

    DashboardStats:
      type: object
      required: [totalDocuments, documentsThisMonth, documentsByType, planLimits]
      properties:
        totalDocuments:
          type: integer
        documentsThisMonth:
          type: integer
        documentsByType:
          type: array
          items:
            type: object
            required: [type, count]
            properties:
              type:
                type: string
              count:
                type: integer
        planLimits:
          type: object
          required: [plan, limit, used]
          properties:
            plan:
              type: string
            limit:
              type: ["integer", "null"]
            used:
              type: integer

    Contract:
      type: object
      required: [id, userId, title, fileName, fileType, fileSize, status, uploadedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        content:
          type: ["string", "null"]
        status:
          type: string
          enum: [processing, ready, error]
        uploadedAt:
          type: string
        processedAt:
          type: ["string", "null"]

    ContractUploadResponse:
      type: object
      required: [contractId, status]
      properties:
        contractId:
          type: integer
        status:
          type: string
          enum: [processing, ready, error]

    RAGQueryInput:
      type: object
      required: [question]
      properties:
        question:
          type: string

    RAGQueryResponse:
      type: object
      required: [answer]
      properties:
        answer:
          type: string