# lib/api-zod/src/generated/types/contract.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LexAI — Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractStatus } from './contractStatus';

export interface Contract {
  id: number;
  userId: number;
  title: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  /** @nullable */
  content?: string | null;
  status: ContractStatus;
  uploadedAt: string;
  /** @nullable */
  processedAt?: string | null;
}