# lib/api-zod/src/generated/types/contractUploadResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LexAI — Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractUploadResponseStatus } from './contractUploadResponseStatus';

export interface ContractUploadResponse {
  contractId: number;
  status: ContractUploadResponseStatus;
}