# lib/api-zod/src/generated/types/rAGQueryResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LexAI — Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface RAGQueryResponse {
  answer: string;
}